+++ src/main/java/xz/fzu/dao/IResumeDao.java
@@ -52,7 +52,7 @@ public interface IResumeDao  extends Mapper<Resume> {
      * @author Murphy
      * @date 2019/4/27 22:52
      */
-    Resume selectInstanceByResumeId(int resumeId);
+    Resume selectInstanceByResumeId(Long resumeId);
 
     /**
      * 查看用户的所有实例
